springbatch/src/main/java/be/cegeka/batchers/springbatch/domain/EmployeeRepository.java
package be.cegeka.batchers.springbatch.domain;


import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

@Repository
@Transactional
public class EmployeeRepository {
    @PersistenceContext
    EntityManager entityManager;

    public void save(Employee employee) {
        entityManager.persist(employee);
    }

    public Employee getBy(Long id) {
        return entityManager.find(Employee.class, id);
    }
}
